backend/requirements.txt
# RoadWatch backend dependencies
# Keep this list minimal — no heavy ML packages unless YOLO is actually used.

# ── Web framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
prometheus-fastapi-instrumentator==7.0.0

# ── Database ──────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2
redis[asyncio]==5.0.4

# ── Auth / security ───────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Request parsing ───────────────────────────────────────────────
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── Config ────────────────────────────────────────────────────────
python-dotenv==1.0.1
slowapi
sentry-sdk[fastapi]==1.45.0
firebase-admin>=6.2.0
apscheduler==3.10.4

# ── File / image handling ─────────────────────────────────────────
Pillow==10.4.0
python-magic==0.4.27
aiofiles==23.2.1
fpdf2==2.7.9
numpy<2.0.0
boto3==1.34.111

# ── Optional: uncomment to enable real YOLOv8 inference ───────────
ultralytics==8.2.0
torch==2.3.0
torchvision==0.18.0